KLapse: Fix compilation error by making function inline
After cherry-picking the KLapse commits, I was getting this warning while compilation: -Wunused-function (the warning was for the function mdss_mdp_kcal_store_fb0_ctl). GCC termed this as a fatal error. Since this was just a warning, there had to be a way around it. According to the GCC documentation, "-Wunused-function: Warn whenever a static function is declared but not defined or a non-inline static function is unused. This warning is enabled by -Wall."

mdss_mdp_kcal_store_fb0_ctl() was defined but wasn't being used, thus the warning. As a quick fix, I resorted to making the function inline. The better solution would be to remove the function and try compiling. I will try doing that when I get the time. FOr the time being, I am deploying the quick fix.