kernel: Change tick rate to 300 HZ
When releasing v6, I wrote this:
The timer frequency -- defined in Hz -- determines the frequency at which the timer interrupt is fired. An interrupt causes the CPU to pause the execution of the current process and allocate its resources to the process or hardware that issued the interrupt. For example, when you press the volume key, an interrupt is fired by the volume keys and the CPU then changes the volume level and shows the changed volume level on the screen. Similarly, interrupts are issued when you touch on the screen or press any hardware key. There is a time gap between two successive interrupts. We call this as the timer frequency. A timer frequency of 100Hz means an interrupt will be issued every 10 milliseconds and a timer frequency of 1000Hz means the interrupt will be fired after 1 millisecond. Which is better? 1000Hz will provide the least latency and the smoothest experience. But it will increase the overhead on the CPU and increase battery drain. 100Hz will keep the CPU overhead minimal but may increase latency. Elindir v6 decreases the timer frequency from 250Hz (interrupts are issued every 4ms) to 100Hz (interrupts are issued every 10ms). In my little testing, I didn't notice any lags that were caused by this change but I did feel the battery lasted slightly longer than usual. If you play games, let me know your feedback on this change.

Several users reported stutters in gameplay with v6 which was resolved after restoring the default value of the timer frequency (also known as the tick rate). Some users noticed improvments with 300Hz, but there weren't any noticeable improvmenets with higher tick rate values like 600Hz. So, sticking to 300Hz for now.